move setup stream fm from root to utils
