+++ src/routes/layout.tsx
@@ -15,7 +15,8 @@ import {
   type ModalStore,
   ModalStoreContext,
 } from "~/interface/web3modal/ModalStore";
-import { setupStream } from "~/root";
+import { setupStream } from "~/utils/stream";
+
 
 const metadata = {
   name: "Web3Modal",
